moved ApiError to model package
